refactor users-service
